requirements.txt
appnope==0.1.3
asttokens==2.4.1
backcall==0.2.0
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
wcwidth==0.2.12
zipp==3.17.0
